+++ RespectU/Sources/Extension/MissionSection.CE+UI.swift
@@ -10,11 +10,13 @@ import UIKit
 
 extension MissionSection.CE {
   
-  /// Creates `CAGradientLayer` matching each mission section.
-  ///
-  /// - Parameter direction: The direction of gradient layer.
-  ///
-  /// - Returns: The created `CAGradientLayer`.
+  var color: UIColor? {
+    switch self {
+    case .electronicCity: return .electronicCity
+    case .metropolis: return .metropolis
+    }
+  }
+  
   func makeGradient(by direction: GradientDirection) -> CAGradientLayer? {
     let startPoint = direction.startPoint
     let endPoint = direction.endPoint
