+++ RespectU/Sources/Extension/MissionSection.LinkDisk+UI.swift
@@ -10,11 +10,13 @@ import UIKit
 
 extension MissionSection.LinkDisk {
   
-  /// Creates `CAGradientLayer` matching each mission section.
-  ///
-  /// - Parameter direction: The direction of gradient layer.
-  ///
-  /// - Returns: The created `CAGradientLayer`.
+  var color: UIColor? {
+    switch self {
+    case .whiteDisk: return .whiteDisk
+    case .blackDisk: return .blackDisk
+    }
+  }
+  
   func makeGradient(by direction: GradientDirection) -> CAGradientLayer? {
     let startPoint = direction.startPoint
     let endPoint = direction.endPoint
