+++ RespectU/Sources/Extension/MissionSection.Respect+UI.swift
@@ -10,11 +10,21 @@ import UIKit
 
 extension MissionSection.Respect {
   
-  /// Creates `CAGradientLayer` matching each mission section.
-  ///
-  /// - Parameter direction: The direction of gradient layer.
-  ///
-  /// - Returns: The created `CAGradientLayer`.
+  var color: UIColor? {
+    switch self {
+    case .departure: return .departure
+    case .clubRoad645: return .clubRoad645
+    case .maxTheater: return .maxTheater
+    case .anotherWorld: return .anotherWorld
+    case .backStage: return .backStage
+    case .chaosTheory: return .chaosTheory
+    case .soundLab: return .soundLab
+    case .visualizer: return .visualizer
+    case .developers: return .developers
+    case .destination: return .destination
+    }
+  }
+  
   func makeGradient(by direction: GradientDirection) -> CAGradientLayer? {
     let startPoint = direction.startPoint
     let endPoint = direction.endPoint
