+++ RespectU/Sources/Extension/MissionSection.Technika2+UI.swift
@@ -10,11 +10,13 @@ import UIKit
 
 extension MissionSection.Technika2 {
   
-  /// Creates `CAGradientLayer` matching each mission section.
-  ///
-  /// - Parameter direction: The direction of gradient layer.
-  ///
-  /// - Returns: The created `CAGradientLayer`.
+  var color: UIColor? {
+    switch self {
+    case .starMixing: return .starMixing
+    case .clubMixing: return .clubMixing
+    }
+  }
+  
   func makeGradient(by direction: GradientDirection) -> CAGradientLayer? {
     let startPoint = direction.startPoint
     let endPoint = direction.endPoint
