--- RespectU/Sources/Extension/MissionSection.Technika3+Gradient.swift
@@ -1,28 +1,0 @@
-//
-//  MissionSection.Technika3+Gradient.swift
-//  RespectU
-//
-//  Created by Presto on 10/03/2019.
-//  Copyright © 2019 Presto. All rights reserved.
-//
-
-import UIKit
-
-extension MissionSection.Technika3 {
-  
-  /// Creates `CAGradientLayer` matching each mission section.
-  ///
-  /// - Parameter direction: The direction of gradient layer.
-  ///
-  /// - Returns: The created `CAGradientLayer`.
-  func makeGradient(by direction: GradientDirection) -> CAGradientLayer? {
-    let startPoint = direction.startPoint
-    let endPoint = direction.endPoint
-    switch self {
-    case .popMixing:
-      return nil
-    case .crewChallenge:
-      return nil
-    }
-  }
-}
