+++ RespectU/Sources/ViewControllers/Top50/Top505BTableViewController.swift
@@ -28,7 +28,7 @@ extension Top505BTableViewController {
     let cell = tableView.dequeueReusableCell(withIdentifier: cellIdentifier, for: indexPath)
     if case let top50Cell as Top50Cell = cell {
       let object = results[indexPath.row]
-      top50Cell.configure(with: object, button: .button8)
+      top50Cell.configure(with: object, button: .button5)
     }
     return cell
   }
