+++ RespectU/Supports/Info.plist
@@ -28,7 +28,7 @@
 		</dict>
 	</array>
 	<key>CFBundleVersion</key>
-	<string>1</string>
+	<string>2</string>
 	<key>FacebookAppID</key>
 	<string>266597110540044</string>
 	<key>FacebookDisplayName</key>
